refactor(desktop): the row anchors leave the search that reads them
`searchAnchorProps` and the ids the non-setting rows wear lived in the search
module, so every file that draws a row imported the search — and the search now
imports the connection table those rows are drawn from. `settings-anchors.ts`
is the anchors alone, which is what breaks the ring: nothing under
`renderer/` imports itself round any more.
