+++ apps/desktop/src/renderer/settings/workspace-rows.tsx
@@ -9,7 +9,7 @@ import { APP_SETTING_SCHEMA } from "@sidecar/settings";
 import type { AppSettingsView } from "@sidecar/settings/wire";
 import type { ActionResult } from "@sidecar/wire";
 import { APP_SETTING_ID } from "../luke-guide";
-import { defaultProjectRowId } from "../settings-search";
+import { defaultProjectRowId } from "../settings-anchors";
 import type { SupersetControl, WorkspaceProviderOption } from "./controls";
 import { SelectRow } from "./select-row";
 import { useSettingWrite } from "./use-setting-write";
